refactor: AppToken과 RefreshToken AuthTokenInfo로 이동 (#325)
