+++ src/main/java/org/sopt/app/application/auth/JwtTokenService.java
@@ -14,8 +14,8 @@ import javax.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.val;
 import org.joda.time.LocalDateTime;
+import org.sopt.app.application.auth.dto.PlaygroundAuthTokenInfo.AppToken;
 import org.sopt.app.application.auth.dto.PlaygroundProfileInfo;
-import org.sopt.app.application.auth.dto.PlaygroundProfileInfo.AppToken;
 import org.sopt.app.application.user.UserInfo;
 import org.sopt.app.common.exception.UnauthorizedException;
 import org.springframework.beans.factory.annotation.Value;
@@ -40,7 +40,7 @@ public class JwtTokenService {
         return Keys.hmacShaKeyFor(Decoders.BASE64.decode(secretKey));
     }
 
-    public PlaygroundProfileInfo.AppToken issueNewTokens(UserInfo.Id userId,
+    public AppToken issueNewTokens(UserInfo.Id userId,
             PlaygroundProfileInfo.PlaygroundMain playgroundMember) {
         val accessToken = this.encodeJwtToken(userId, playgroundMember.getId());
         val refreshToken = this.encodeJwtRefreshToken(userId);
