+++ src/test/java/org/sopt/app/facade/AuthFacadeTest.java
@@ -11,9 +11,9 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.sopt.app.application.auth.JwtTokenService;
-import org.sopt.app.application.auth.dto.PlaygroundProfileInfo.AppToken;
+import org.sopt.app.application.auth.dto.PlaygroundAuthTokenInfo.AppToken;
+import org.sopt.app.application.auth.dto.PlaygroundAuthTokenInfo.RefreshedToken;
 import org.sopt.app.application.auth.dto.PlaygroundProfileInfo.PlaygroundMain;
-import org.sopt.app.application.auth.dto.PlaygroundProfileInfo.RefreshedToken;
 import org.sopt.app.application.auth.PlaygroundAuthService;
 import org.sopt.app.application.soptamp.SoptampPointService;
 import org.sopt.app.application.soptamp.SoptampUserService;
